+++ tests/reorder_code/expected/reorder_typed_array_cast.gd
@@ -4,4 +4,8 @@ static func make_defs() -> Variant:
 		SomeType.new(&"bbb"),
 		SomeType.new(&"ccc"),
 		SomeType.new(&"ddd"),
+		SomeType.new(&"eee"),
+		SomeType.new(&"fff"),
+		SomeType.new(&"ggg"),
+		SomeType.new(&"hhh"),
 	] as Array[SomeType]