+++ tests/reorder_code/input/reorder_typed_array_cast.gd
@@ -1,7 +1,2 @@
 static func make_defs() -> Variant:
-	return [
-		SomeType.new(&"aaa"),
-		SomeType.new(&"bbb"),
-		SomeType.new(&"ccc"),
-		SomeType.new(&"ddd"),
-	] as Array[SomeType]
+	return [SomeType.new(&"aaa"), SomeType.new(&"bbb"), SomeType.new(&"ccc"), SomeType.new(&"ddd"), SomeType.new(&"eee"), SomeType.new(&"fff"), SomeType.new(&"ggg"), SomeType.new(&"hhh")] as Array[SomeType]