+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This file documents the changes made to the formatter with each release.
 
 - Handle Tree-sitter parse errors gracefully, preserve declarations with parse errors while formatting valid code around them (#279)
 
+### Changed
+
+- Force short lines that were manually wrapped on multiple lines with commas to merge back into one line when they fit within the max line length (this behavior was a leftover from the old formatter implementation)
+
 ### Fixed
 
 - Fix long ternaries without pre-existing backslashes or delimiters formatting in a way that GDScript doesn't parse (the formatter will now add parentheses around the expression and indent it) (#293)